Make result depend on algorithm data only
