+++ frontend/src/Components/Result/ResultChart.css
